+++ .prettierrc.js
@@ -2,6 +2,7 @@ module.exports = {
   singleQuote: true,
   tabWidth: 2,
   printWidth: 120,
+  semi: true,
   overrides: [
     {
       files: '*.sol',
