+++ client/.env.example
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_INFURA_ID=
