+++ client/.prettierrc.js
@@ -1,0 +1,10 @@
+module.exports = {
+	semi: true,
+	singleQuote: true,
+	printWidth: 120,
+	tabWidth: 4,
+	trailingComma: 'es5',
+	useTabs: true,
+	bracketSpacing: true,
+	arrowParens: 'avoid',
+}
