+++ client/package.json
@@ -1,32 +1,35 @@
 {
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "pkg:update:check": "ncu",
-    "pkg:update:install": "ncu -u; npm install",
-    "pkg:update:int": "ncu -i"
-  },
-  "dependencies": {
-    "@headlessui/react": "1.5.0",
-    "axios": "0.26.1",
-    "ethers": "5.6.0",
-    "install": "0.13.0",
-    "next": "12.1.0",
-    "next-themes": "0.1.1",
-    "npm": "8.5.4",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-loader-spinner": "5.1.4",
-    "wagmi": "0.2.18",
-    "zksync": "0.11.6"
-  },
-  "devDependencies": {
-    "autoprefixer": "10.4.2",
-    "daisyui": "2.6.4",
-    "postcss": "8.4.8",
-    "tailwindcss": "3.0.23"
-  }
+	"name": "wallet-dashboard",
+	"version": "0.1.0",
+	"private": true,
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint"
+	},
+	"dependencies": {
+		"@headlessui/react": "1.6.6",
+		"@heroicons/react": "1.0.6",
+		"@rainbow-me/rainbowkit": "0.4.4",
+		"autoprefixer": "10.4.7",
+		"axios": "0.27.2",
+		"daisyui": "2.20.0",
+		"ethers": "5.6.9",
+		"next": "12.2.3",
+		"next-themes": "0.2.0",
+		"postcss": "8.4.14",
+		"react": "18.2.0",
+		"react-dom": "18.2.0",
+		"react-loader-spinner": "^5.1.5",
+		"tailwindcss": "3.1.6",
+		"wagmi": "0.5.11"
+	},
+	"devDependencies": {
+		"@types/node": "18.0.6",
+		"@types/react": "18.0.15",
+		"eslint": "8.20.0",
+		"eslint-config-next": "12.2.3",
+		"typescript": "4.7.4"
+	}
 }
