+++ client/src/lib/consts.ts
@@ -1,0 +1,1 @@
+export const APP_NAME = 'Wallet Dashboard'
