+++ client/src/lib/shortenAddress.js
