+++ client/src/pages/api/hello.ts
@@ -1,0 +1,7 @@
+import { NextApiRequest, NextApiResponse } from 'next'
+
+const handler = (req: NextApiRequest, res: NextApiResponse) => {
+	res.status(200).json({ name: 'hello, world!' })
+}
+
+export default handler
