+++ client/tailwind.config.js
@@ -1,13 +1,11 @@
+/** @type {import('tailwindcss/tailwind-config').TailwindConfig} */
 module.exports = {
-  content: [
-    "./pages/**/*.{js,ts,jsx,tsx}",
-    "./components/**/*.{js,ts,jsx,tsx}",
-  ],
-  theme: {
-    extend: {},
-  },
-  daisyui: {
-    themes: ["light", "dark"],
-  },
-  plugins: [require("daisyui")],
-};
+	content: ['./src/**/*.{js,ts,jsx,tsx}'],
+	theme: {
+		extend: {},
+	},
+	daisyui: {
+		themes: ['light', 'dark'],
+	},
+	plugins: [require('daisyui')],
+}
