+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "lsbdev-hardhat-contracts",
+  "name": "wallet-dashboard",
   "scripts": {
     "runContract": "npx hardhat run scripts/run.js",
     "deployLocal": "npx hardhat run scripts/deploy.js --network localhost",
@@ -13,16 +13,16 @@
     "pkg:update:int": "ncu -i"
   },
   "devDependencies": {
-    "@nomiclabs/hardhat-ethers": "2.0.6",
+    "@nomiclabs/hardhat-ethers": "2.1.0",
     "@nomiclabs/hardhat-etherscan": "3.1.0",
     "@nomiclabs/hardhat-waffle": "2.0.3",
-    "@openzeppelin/contracts": "4.6.0",
+    "@openzeppelin/contracts": "4.7.1",
     "chai": "4.3.6",
     "dotenv": "16.0.1",
     "ethereum-waffle": "3.4.4",
-    "ethers": "5.6.8",
-    "hardhat": "2.9.7",
-    "typescript": "4.7.3",
-    "zksync-web3": "0.4.0"
+    "ethers": "5.6.9",
+    "hardhat": "2.10.1",
+    "typescript": "4.7.4",
+    "zksync-web3": "0.7.9"
   }
 }
